Move wrapper class from bind.py to wrapper.py
